% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hants.R
\name{hants}
\alias{hants}
\title{Harmonic analysis for time series: fitting the harmonic regression model under the presence
of outliers}
\usage{
hants(y, ts = 1:length(y), lenBasePeriod = length(y), numFreq,
  HiLo = c("Hi", "Lo"), low, high, fitErrorTol, degreeOverDeter, delta)
}
\arguments{
\item{y}{numeric vector containing time series on which harmonic 
regression will be fitted. Missing values are not allowed}

\item{ts}{numeric vector of length \code{y} with the sampling
points for \code{y}. Default is \eqn{ts[i] = i, i=1,\ldots, 
\code{y}}}

\item{lenBasePeriod}{numeric giving the length of the base period, reported in 
samples, e.g. days, dekads, months, years, etc.}

\item{numFreq}{numeric indicating the total number of frequencies to be 
used in the harmonic regression}

\item{HiLo}{character indicating whether high or low outliers must be rejected}

\item{low}{numeric giving minimum valid value of fitted harmonic regression 
function}

\item{high}{numeric giving maximum valid value of fitted harmonic regression
function}

\item{fitErrorTol}{numeric giving maximum allowed distance between observations and fitted
curve; if difference between a given observation and its fitted value 
exceeds \code{fitErrorTol} then the observation is excluded from
next iteration of fitting algorithm}

\item{degreeOverDeter}{numeric; iteration stops when number of observations equals
number of observations for curve fitting plus \code{degreeOverDeter};
the latter in turns is by definition \code{y} minus
\eqn{min(2 * \code{numFreq} +1, \code{y})}.}

\item{delta}{numeric giving a (small) regularization parameter to prevent 
non-invertible hat matrix (see \code{details}), probably caused by high
amplitudes.}
}
\value{
A list containing:
\item{amplitude}{a numeric vector with amplitude estimates.}
\item{phase}{a numeric vector with phase estimates.}
\item{fitted}{a numeric vector with fitted values via harmonic regression.}
}
\description{
iterative algorithm that computes amplitudes and phase angles in the 
typical harmonic regression framework. Based on these estimates a harmonic 
regression function is fitted. As part of the iterative algorithm, observations
are being excluded from the design matrix of the regression model if the distance
between them and the fitted curve exceeds the value of the parameter \code{fitErrorTol}.
\code{hants} is based on implementations of the same name written in Fortran 
and Matlab computer languages.
}
\details{
This function does not allow missing data, missing values of the original
data \code{y} must be substituted by values considerably out of the range of the observations.
}
\examples{
y <- c(5, 2, 5, 10, 12, 18, 20, 23, 27, 30, 40, 60, 66, 
      70, 90, 120, 160, 190, 105, 210, 104, 200, 90, 170, 
      50, 120, 80, 60, 50, 40, 30, 28, 24, 20, 15, 10)
fitLow <- hants(y = y, numFreq = 3, HiLo = "Lo", low = 0, 
          high = 255, fitErrorTol = 5, degreeOverDeter = 1, delta = 0.1)
fitHigh <- hants(y = y, numFreq = 3, HiLo = "Hi", low = 0, 
          high = 255, fitErrorTol = 5, degreeOverDeter = 1, delta = 0.1)
fit_harmR <- harmonicR(y = y, numFreq = 5, delta = 0.1)
plot(y, pch = 16, main = "hants fitting")
lines(fitLow$fitted, lty = 4, col = "red")
lines(fitHigh$fitted, lty = 2, col = "blue")
lines(fit_harmR$fitted, lty = 4, col = "green")

# A missing value should be substituted by a number out of the valid range
y1 <- y
y1[20] <- -10
fitLowMiss <- hants(y = y1, numFreq = 3, HiLo = "Lo", low = 0, high = 255, 
              fitErrorTol = 5, degreeOverDeter = 1, delta = 0.1)
fitHighMiss <- hants(y = y1, numFreq = 3, HiLo = "Hi", low = 0, 
               high = 255, fitErrorTol = 5, degreeOverDeter = 1, delta = 0.1)
fit_harmRMiss <- harmonicR(y = y1, numFreq = 3, delta = 0.1)
plot(y1, pch = 16, main = "hants fitting (missing values)")
lines(fitLowMiss$fitted, lty = 4, col = "red")
lines(fitHighMiss$fitted, lty = 2, col = "blue")
lines(fit_harmRMiss$fitted, lty = 4, col = "green")

}
\references{
Roerink, G.J., Menenti, M., Verhoef, W. (2000).
\emph{Reconstructing clodfree NDVI composites using Fourier analysis of time series}, 
Int. J. Remote Sensing, \bold{21(9)}, 1911--1917.

The Matlab implementation of HANTS can be found 
\href{https://nl.mathworks.com/matlabcentral/fileexchange/38841-matlab-implementation-of-harmonic-analysis-of-time-series-hants}{here}.
}
\seealso{
\code{\link[geoTS]{harmonicR}}
}
