% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicR.R
\name{harmonicR}
\alias{harmonicR}
\title{Harmonic regression model}
\usage{
harmonicR(y, lenBasePeriod = length(y), numFreq, delta)
}
\arguments{
\item{y}{numeric vector containing time series on which harmonic 
regression will be fitted; missing values are not allowed}

\item{lenBasePeriod}{numeric}

\item{numFreq}{numeric indicating the total number of frequencies to be 
used in the harmonic regression.}

\item{delta}{numeric giving a (small) regularization parameter to prevent 
non-invertible hat matrix (see \code{details}), probably caused by high
amplitudes.}
}
\value{
A list containing:
\item{amplitude}{a numeric vector with amplitude estimates.}
\item{phase}{a numeric vector with phase estimates.}
\item{fitted}{a numeric vector with fitted values via harmonic regression.}
}
\description{
Computes amplitudes and phase angle in the typical harmonic regression framework.
Based on these estimates a harmonic regression function is fitted.
}
\details{
This function does not allow missing data, missing values of the original
data \code{y} must be substituted by values considerably out of the range of the observations.
}
\examples{
y <- c(5, 2, 5, 10, 12, 18, 20, 23, 27, 30, 40, 60, 66, 70, 90, 120, 
     160, 190, 105, 210, 104, 200, 90, 170, 50, 120, 80, 60, 50, 40, 
     30, 28, 24, 20, 15, 10)
y2 <- y[-c(19, 21, 23, 25)]
fit <- harmonicR(y = y2, numFreq = 5, delta = 0.01)
(fit)
plot(y2, pch = 16, col = "lightgray")
lines(fit$fitted)

y3 <- y[-c(16:18, 20, 22, 24, 26)]
fit1 <- harmonicR(y = y3, numFreq = 5, delta = 0.01)
fit2 <- harmonicR(y = y3, numFreq = 8, delta = 0.01)
fit3 <- harmonicR(y = y3, numFreq = 10, delta = 0.01)
plot(y3, pch = 16, col = "lightgray")
lines(fit1$fitted)
lines(fit2$fitted, lty = 4)
lines(fit3$fitted, lty = 2)

}
\seealso{
\code{\link[geoTS]{hants}}
}
