% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{master}
\alias{master}
\title{Land Mask of eastern Yucatan Peninsula}
\description{
A \code{RasterLayer} with a spatial subset covering eastern Yucatan Peninsula
of Mexico. A land mask is a binary layer where \code{1=Land}, \code{0=Water}.
}
\section{master.tif}{


A \code{RasterLayer} object with 500 rows, 600 columns. Each cell has a resolution
of 250m.
}

