% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{MOD13Q1_NDVI_2000129_009}
\alias{MOD13Q1_NDVI_2000129_009}
\title{MOD13Q1 NDVI binary file}
\description{
Spatial subset of a MOD13Q1 NDVI layer, in binary format, covering eastern 
Yucatan Peninsula, Mexico. \href{https://en.wikipedia.org/wiki/Normalized_difference_vegetation_index}{NDVI} stands for Normalized Difference
Vegetation Index; NDVI = (NIR-RED)/(NIR+RED) where NIR and RED are the Near
Infrared and Red bands of the MODIS product, respectively. More information
about the MODIS mission can be found \href{https://modis.gsfc.nasa.gov/about/}{here}.
}
\section{MOD13Q1_NDVI_2000129_009.bin}{


This image contains NDVI measurements for the 129-th Julian day of 2000; the
subscript \emph{_009} signifies that this was the 9th NDVI observation of 2000.
}

