% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{MOD13Q1_NDVI_Mohinora}
\alias{MOD13Q1_NDVI_Mohinora}
\title{16-day MOD13Q1 NDVI RasterStack}
\description{
A \code{RasterStack} containing 23 layers of NDVI for 2001. This \code{RasterStack} 
is a spatial subset covering the Protected Area of Flora and Fauna \bold{Cerro Mohinora} 
at Chihuahua, Mexico.
}
\section{MOD13Q1_NDVI_Mohinora.tif}{


A \code{RasterStack} object with 59 rows, 93 columns, 5487 cells and 23 layers.
}

