% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_mohinora}
\alias{shp_mohinora}
\title{SpatialPolygonsDataFrame for Cerro Mohinora}
\format{
An object of class \code{SpatialPolygonsDataFrame}.
}
\usage{
data(shp_mohinora)
}
\description{
A \code{RData} file containing a \code{SpatialPolygonsDataFrame} object
delimiting \bold{Cerro Mohinora} at Chihuhua, the smallest Protected Area 
of Flora and Fauna in Mexico.
}
\keyword{datasets}
