% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_replace_raster.R
\name{split_replace_raster}
\alias{split_replace_raster}
\title{Split a Raster* object and replace cell values (optional)}
\usage{
split_replace_raster(raster, partPerSide, save = T, replace = F,
  valToReplace, replacedBy, dataType, cellsToProcess = 1:(partPerSide^2),
  format = "GTiff", outputPath, name, ...)
}
\arguments{
\item{raster}{Raster* object}

\item{partPerSide}{numeric indicating the number of cells in which \code{raster} will be split}

\item{save}{logical, should the output be saved, default is \code{TRUE}}

\item{replace}{logical, default \code{FALSE}, when \code{TRUE}, \code{valToReplace} and \code{replacedBy} must by specified}

\item{valToReplace}{indicates a value to be replaced across \code{raster} cells}

\item{replacedBy}{indicates the value by which \code{valToReplace} is replaced}

\item{dataType}{character, output data type. See \code{\link[raster]{dataType}}}

\item{cellsToProcess}{numeric vector indicating which of the \code{partPerSide^2} should be processed/saved}

\item{format}{character, output file type, default \code{"GTiff"}. See \code{\link[raster]{writeFormats}}}

\item{outputPath}{character with full path name where the resulting Raster* objects will be saved}

\item{name}{character with the name to assign to final products}

\item{...}{additional arguments used by \code{\link[raster]{writeRaster}}}
}
\value{
At \code{outputPath} the user will find \code{length(cellsToProcess)} files
}
\description{
This function will split a Raster* object into \code{partPerSide^2} parts. Additionally,
it allows to replace cell values (\code{valToReplace}) within Raster* object by another
value of user's choice (\code{replacedBy}). When \code{save = T}, the resulting \code{cellsToProcess} 
Raster* objects are saved in directory \code{outputPath}.
}
\details{
Before processing any of the \code{cellsToProcess} the temporary raster 
directory is re-directed. Basically, prior to process the i-th cell, 
at \code{outputPath} a new subdirectory is created, which, in turn, is erased 
automatically once the i-th cell has been processed. As a result of multiple testing 
we found that this measure avoids memory overflow.
}
\seealso{
\code{\link[raster]{writeRaster}}, \code{\link[raster]{aggregate}}, 
\code{\link[raster]{rasterOptions}}
}
