% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_bin_raster.R
\name{transfer_bin_raster}
\alias{transfer_bin_raster}
\title{Transfer values of a binary into a raster file}
\usage{
transfer_bin_raster(inputPath, outputPath, masterFile, what = integer(),
  signed = T, endian = "little", size = 2, format = "GTiff",
  dataType = "INT2S", overwrite = T)
}
\arguments{
\item{inputPath}{character with full path name of input file(s). A list of files is also allowed.}

\item{outputPath}{character with full path name (where the raster files will be saved).}

\item{masterFile}{character with full path name of a raster file with basic specifications (such
as extent and projections) to be used in the transfer process.}

\item{what}{See \code{\link[base]{readBin}}. Default \code{integer()}.}

\item{signed}{See \code{\link[base]{readBin}}. Default \code{T}.}

\item{endian}{See \code{\link[base]{readBin}}. Default \code{"little"}.}

\item{size}{integer, number of bytes per element in the byte stream, default 2. See \code{\link[base]{readBin}}}

\item{format}{character, output file type. See \code{\link[raster]{writeFormats}}.}

\item{dataType}{character, output data type. See \code{\link[raster]{dataType}}.}

\item{overwrite}{logical, default \code{T}, should the resulting raster be overwritten.}
}
\value{
At the designated path (\code{outputFile}) the user will find an \code{RData} file
}
\description{
Get the values of a binary file and transfer them into a raster file. All formats
considered in \code{\link[raster]{writeRaster}} are allowed.
}
