% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixToRaster.R
\name{matrixToRaster}
\alias{matrixToRaster}
\title{Creates a RasterLayer object from a matrix}
\usage{
matrixToRaster(matrix, RASTER)
}
\arguments{
\item{matrix}{a matrix object.}

\item{RASTER}{a RasterLayer object whose extent and projection will be used to
create a raster from \code{matrix}.}
}
\value{
A RasterLayer
}
\description{
Transforms a \code{matrix} into a \code{RasterLayer} object.
}
\details{
The \code{\link[sp]{coordinates}} and \code{\link[raster]{projection}} of 
the argument \code{RASTER} are used to create a raster from the argument \code{matrix}.
}
